requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
python-dateutil==2.8.2
PyJWT==2.8.0
Werkzeug==2.3.7 